requirements.txt
python-mpv>=1.0.1
PySide6>=6.3.0
